feat: CUD 책임 EntityPersister로 이전
