+++ README.md
@@ -7,3 +7,6 @@
 - [x] entity persister를 통한 update 쿼리 실행
 - [x] entity persister를 통한 insert 쿼리 실행
 - [x] entity persister를 통한 delete 쿼리 실행
+
+## 요구사항 2 - EntityManager의 책임 줄여주기
+- [x] EntityManager의 persist, remove 책임 EntityPersister로 이동
