+++ src/main/java/persistence/entity/EntityPersister.java
@@ -1,33 +1,10 @@
 package persistence.entity;
 
-import jdbc.JdbcTemplate;
-import persistence.sql.dml.DeleteQueryBuilder;
-import persistence.sql.dml.InsertQueryBuilder;
-import persistence.sql.dml.UpdateQueryBuilder;
+public interface EntityPersister {
 
-public class EntityPersister {
+    boolean update(Object entity);
 
-    private final JdbcTemplate jdbcTemplate;
+    void insert(Object entity);
 
-    public EntityPersister(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public boolean update(Object entity) {
-        UpdateQueryBuilder updateQueryBuilder = new UpdateQueryBuilder();
-        String query = updateQueryBuilder.build(entity);
-        return jdbcTemplate.executeForUpdate(query);
-    }
-
-    public void insert(Object entity) {
-        InsertQueryBuilder insertQueryBuilder = new InsertQueryBuilder();
-        String query = insertQueryBuilder.build(entity);
-        jdbcTemplate.execute(query);
-    }
-
-    public void delete(Object entity) {
-        DeleteQueryBuilder deleteQueryBuilder = new DeleteQueryBuilder();
-        String query = deleteQueryBuilder.build(entity);
-        jdbcTemplate.execute(query);
-    }
+    void delete(Object entity);
 }
